event package class tests moved to proper location
