+++ nostr-java-event/build.gradle
@@ -7,3 +7,8 @@ description = 'nostr-java-event'
 dependencies {
     api project(':nostr-java-base')
 }
+
+tasks.named('test') {
+    systemProperty("spring.profiles.active", "test")
+    useJUnitPlatform()
+}
