+++ nostr-java-event/src/test/resources/application-test.properties
@@ -1,0 +1,4 @@
+spring.threads.virtual.enabled=true
+
+logging.level.nostr.base=INFO
+logging.pattern.console=%msg%n
