+++ nostr-java-event/src/test/resources/junit-platform.properties
@@ -1,0 +1,7 @@
+# junit-platform.properties
+
+junit.jupiter.execution.parallel.enabled=true
+junit.jupiter.execution.parallel.config.strategy=dynamic
+junit.jupiter.execution.parallel.mode.default=same_thread
+#junit.jupiter.execution.parallel.mode.default=concurrent
+#junit.jupiter.execution.parallel.mode.classes.default=concurrent
