--- nostr-java-test/src/test/java/nostr/test/event/BaseTagTest.java
@@ -1,29 +1,0 @@
-package nostr.test.event;
-
-import nostr.event.BaseTag;
-import nostr.event.impl.GenericTag;
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-class BaseTagTest {
-
-    BaseTag genericTag = GenericTag.create("id", 1, "value");
-
-    @Test
-    void getNip() {
-        assertEquals(1, genericTag.getNip());
-    }
-
-    @Test
-    void testHashCode() {
-        assertEquals(112174237, genericTag.hashCode());
-    }
-
-    @Test
-    void testToString() {
-        String result = "GenericTag(code=id, nip=1, attributes=[ElementAttribute(name=param0, value=value, nip=null)])";
-        assertEquals(result, genericTag.toString());
-    }
-
-}
